+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Bug #30: Update expected type for attribute value parameters in multiple test cases (@terabytesoftw)
 - Enh #31: Add `HasStrokeOpacity` trait and corresponding tests for managing SVG `stroke-opacity` attribute (@terabytesoftw)
 - Bug #32: Update validation for `stroke-miterlimit` to require a minimum value of `1` (@terabytesoftw)
+- Bug #33: Refactor `fill-opacity` and `opacity` handling with type validation and new tests for SVG attributes (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024